Session_And_Task_Management_Service/Session_And_Task_Management/requirements.txt
amqp==5.3.1
asgiref==3.8.1
autopep8==2.3.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.42.0
confluent-kafka==2.8.0
Django==5.1.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.5.0
django-environ==0.12.0
django-silk==5.3.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gprof2dot==2024.6.6
idna==3.10
kombu==5.4.2
loadenv==0.1.1
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pycodestyle==2.12.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
boto3==1.37.16
botocore==1.37.16
django-storages==1.14.5
jmespath==1.0.1
s3transfer==0.11.4
Pillow==11.1.0
olefile==0.47
gunicorn==23.0.0
django-structlog==9.1.1
python-json-logger==3.3.0
prometheus-client==0.21.1
structlog==25.2.0